AI Switch logic improvements and optimizations
- **More AI switching logic improvements**
  - Optimized the mid-turn and Post-KO switching algorithms which resulted in decent performance gains in simulation times with medium and hard difficulty AIs.
  - Better AI switching logic. Should result in much less switch jittering, more consistent and conservative decision making.
    - AI switching won't be as aggressive to switch to another Pokemon.
    - AI shouldn't be sending in a Pokemon to its death.
- **Simulations now added into the main menu**
  - You can sim x amount of iterations of 2 customizable AI teams. You can use this for benchmarking and/or see how different trainers at different AI levels compare against eachother.
  - **NOTE:** The simulations will utilize all your CPU cores to the fullest. Recommend not to go over 10 million iterations. Even less iterations on less powerful CPUs. It's just gonna take a while.
    - As an example on my 8 core 16 thread AMD 5800x CPU, Red trainer vs Blue trainer and with both AIs at hard difficulty, 10 million iterations takes around 25-27 seconds.
- **Developer notes**
  - Besides the AI switching algorithms, I also converted my Move Routines to use compile-time function pointer jump table instead std::variant which resulted in a negligible speed bump, but slightly cleaner code.
  - The other big performance gain was actually from making sure to not needlessly run RandomizePostTurnOrder() in my PostTurnEffectProcessor after every status check.
    - The RandomizePostTurnOrder() method only runs when both Pokemon have a specific status (such as both having leech seed or both having some sort of damage status or same field effect).
  - Stage ratios are no longer stored, they are calculated but result in the same outcomes. This allowed for multiple files having access to the same GetStageRatio() constexpr functions.
  - Added MoveRoutines namespace. 
  - StageRatio.h, AppState.h, PokedexConstexpr.h, MovedexConstexpr.h, and LearnsetData.h files added.
    - Removed LoadEmbed .cpp files.
  - Instead of Pokemon, moves and learnsets being instantiated via std::vector, they're compile-time constant arrays now.
  - Experimented with data types. A lot of int types are now unsigned int types. Also using size_t a little more when indexing into contiguous memory data structures.

+++ src/moves/MoveHelpers.h
@@ -17,23 +17,23 @@ void DamageRoutine(MoveRoutineDeps&);
 
 void MultiStrikeRoutine(MoveRoutineDeps&, int turnCount);
 
-void FixedDamageRoutine(MoveRoutineDeps&, int baseDamage);
+void FixedDamageRoutine(MoveRoutineDeps&, unsigned int baseDamage);
 
 void FlinchRoutine(MoveRoutineDeps&);
 
-void RecoilRoutine(MoveRoutineDeps& deps, int recoilDivisor, int targetHPBegin, int targetHPEnd);
+void RecoilRoutine(MoveRoutineDeps& deps, unsigned int recoilDivisor, unsigned int targetHPBegin, unsigned int targetHPEnd);
 
-using GetStageFn = int (*)(BattlePokemon&);
-using SetStageFn = void (*)(BattlePokemon&, int);
+using GetStageFn = size_t (*)(BattlePokemon&);
+using SetStageFn = void (*)(BattlePokemon&, size_t);
 
-void StageUpRoutine(MoveRoutineDeps& deps, int amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
-void StageDownRoutine(MoveRoutineDeps& deps, int amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
-void StageDownDamageRoutine(MoveRoutineDeps& deps, int amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
-void DisplayStatChange(IMoveResultsUI& ui, int amount, bool isUp, std::string_view stageName, std::string_view playerName, std::string_view pokemonName);
+void StageUpRoutine(MoveRoutineDeps& deps, size_t amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
+void StageDownRoutine(MoveRoutineDeps& deps, size_t amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
+void StageDownDamageRoutine(MoveRoutineDeps& deps, size_t amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
+void DisplayStatChange(IMoveResultsUI& ui, size_t amount, bool isUp, std::string_view stageName, std::string_view playerName, std::string_view pokemonName);
 
 using ChargeMsgMemFn = void (IMoveResultsUI::*)(std::string_view, std::string_view) const;
 using ChargeHookFn = void (*)(MoveRoutineDeps& deps);
-using ChargeStageUpFn = void (*)(MoveRoutineDeps& deps, int amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
+using ChargeStageUpFn = void (*)(MoveRoutineDeps& deps, size_t amount, std::string_view stageName, GetStageFn getStage, SetStageFn setStage);
 
 struct ChargingMoveHooks
 {
@@ -42,7 +42,7 @@ struct ChargingMoveHooks
     ChargeStageUpFn stageUp{ nullptr };
     GetStageFn getStage{ nullptr };
     SetStageFn setStage{ nullptr };
-    int stageIncreaseAmount{};
+    size_t stageIncreaseAmount{};
     std::string_view stageName{};
 };
